feat: modularize portfolio hooks
